Add get_available_ports and get_random_tcp_uris
Add `get_available_ports` and `get_random_tcp_uris` functions to return
a list of the specified number of distinct TCP uris, each with a unique
(and currently unused) port number.  This is necessary in the case where
`get_random_tcp_uri` must be called before binding to the previously
returned URI.  Otherwise, it is possible that the same port will be
returned twice, which would lead to errors about the port already being
in use.