[NB] Add UUID attr to proxy classes, add UUID frame...
Add UUID attr to proxy classes, allowing unique referencing of any proxy
instance.

Use the first frame to store the sender UUID of each RPC request.

[NB] This changes the base RPC handler method signature from:

    def handler(self, env, *args, **kwargs):
        ...

to:

    def handler(self, env, sender_uuid, *args, **kwargs):
        ...

The first frame of each _response_ is set to a string of the date and
time the request was completed (not the UUID).