Add deferred command ZeroMQ `__flags__` kwarg
This allows an rpc proxy method to be called with ZeroMQ flags set to
something other than the default.  As an example of when this might be
useful, add `tests/test_rpc_noblock.py`, which shows how to perform
asynchronous processing of an RPC request, by passing
`__flags__=zmq.NOBLOCK` to the deferred command call.  This way, if
there is no response ready, an exception will be raised.  This exception
can be caught and manual calls can be made to the deferred command's
`recv_response` method to try again.