[NB] Allow sets of handlers to be managed...
Allow sets of handlers to be managed, where each set of handlers is
grouped by a common prefix.  This is done by adding a `BaseHandlerMixin`
class, which can be sub-classed to set the list of prefixes to scan for.

The `get_handler` method looks up a handler from the set for the
specified handler prefix.  If no handler prefix is specified, the first
handler prefix in the `handler_prefixes` class attribute will be used.